Relocate dgal into bathmery package, and Relocate cogJob types into shared package.
