+++ packages/bathymetry/src/gdal/gdal.command.ts
