+++ packages/bathymetry/src/gdal/gdal.docker.ts
