+++ packages/bathymetry/src/gdal/gdal.local.ts
