+++ packages/bathymetry/src/gdal/gdal.ts
