+++ packages/cli/package.json
@@ -55,7 +55,6 @@
     "ansi-colors": "^4.1.1",
     "deep-diff": "^1.0.2",
     "flatgeobuf": "^3.23.1",
-    "node-fetch": "^3.2.3",
     "p-limit": "^4.0.0",
     "pretty-json-log": "^1.0.0",
     "slugify": "^1.6.5",
