+++ packages/shared/package.json
@@ -32,6 +32,7 @@
     "@linzjs/metrics": "^6.41.0",
     "aws-sdk": "^2.890.0",
     "entities": "^4.3.0",
+    "node-fetch": "^3.2.3",
     "pino": "^8.6.1",
     "source-map-support": "^0.5.19",
     "ulid": "^2.3.0"
