+++ packages/shared/src/cog/gdal.config.ts
