+++ packages/shared/src/cog/stac.ts
